requirements.txt
boto3
requests
google-auth
google-cloud-storage
urllib3<2.0